requirements.txt
requests
pyyaml
python-telegram-bot